[UI] Phase 5.b.4: Migrate workspace and access modals to shared primitives (#18755)
Migrate the workspace + access dialog flows onto the shared Modal /
InfoModal / FormModal primitives from `@/components/shared/Modal` (#19368)
and relocate the files alongside the surfaces that own them.

Moves
- ui/components/shared/Modal/ViewInfoModal.tsx
    -> ui/components/workspaces/ViewInfoModal.tsx
  Rewritten on the shared `Modal` primitive so the Copy Link / Save
  actions plug into the modal footer instead of an in-body action box.
- ui/components/workspaces/SpacesSwitcher/WorkspaceModal.tsx
    -> ui/components/workspaces/WorkspaceFormModal.tsx
  Rewritten on the shared `Modal` primitive (size="xl"). The drawer +
  content navigation is extracted into WorkspaceFormModalNav.tsx and
  WorkspaceFormModalSections.tsx so each file stays under the project's
  per-file size budget. The dead `StyledModal` styled-component is
  dropped from `SpacesSwitcher/styles.tsx`.
- ui/components/workspaces/SpacesSwitcher/ShareModal.tsx
    -> ui/components/workspaces/ShareWorkspaceModal.tsx
  Relocated; Sistent's `ShareModal` (`CatalogShare`) still owns the
  dialog UI, so the wrapper stays a thin RTK Query adapter rather than
  composing a project-side primitive (which would double-wrap the
  dialog and break sharing behaviour).
- ui/pages/extension/AccessMesheryModal.tsx
    -> ui/components/layout/AccessMesheryModal.tsx
  Rewritten on the shared `Modal` primitive, replacing the bespoke MUI
  `Dialog` + `DialogTitle` + `Slide` chrome.

Consumer updates
- Imports updated in:
  - components/layout/Navigator/NavigatorExtension.tsx
  - components/workspaces/SpacesSwitcher/MainViewsContent.tsx
  - components/workspaces/SpacesSwitcher/MainDesignsContent.tsx
  - components/workspaces/SpacesSwitcher/MyDesignsContent.tsx
  - components/workspaces/SpacesSwitcher/MyViewsContent.tsx
  - components/workspaces/SpacesSwitcher/SpaceSwitcher.tsx
  - pages/_app.tsx

Allowlist housekeeping
- ui/eslint.config.js: removed the deleted paths from
  legacyRestrictedImportOffenders, legacyLiteralColorOffenders,
  legacyMaxLineSoftOffenders so the new files pass the guardrails
  cleanly. No entries added.

Verification
- npx tsc --noEmit: clean
- npx eslint .                   : no new errors/warnings on the
  changed files (pre-existing warnings elsewhere unchanged)
- node scripts/audit-mui.js      : 0 matches (unchanged)
- node scripts/audit-hex.js      : 233 matches in 82 files (was 235 / 83)
- node scripts/audit-size.js     : 8 over_warn / 0 over_hard (unchanged)

Closes #18755
Refs #18654, #18661, #18751
