+++ ui/components/workspaces/ShareWorkspaceModal.tsx
@@ -1,3 +1,14 @@
+/**
+ * Workspace share modal wrapper.
+ *
+ * Thin adapter that wires Meshery's RTK Query hooks (resource access actors,
+ * owner data, visibility mutations) into Sistent's `ShareModal` (aliased here
+ * as `CatalogShare`). The dialog itself is rendered by Sistent, so this
+ * wrapper does not compose a project-side modal primitive — doing so would
+ * double-wrap the dialog and break the sharing behaviour.
+ */
+import { FC } from 'react';
+import { ShareModal as CatalogShare } from '@sistent/sistent';
 import { useUpdatePatternFileMutation } from '@/rtk-query/design';
 import {
   useCreateAndRevokeResourceAccessRecordMutation,
@@ -10,13 +21,30 @@ import {
 } from '@/rtk-query/user';
 import { useUpdateViewVisibilityMutation } from '@/rtk-query/view';
 import { useNotification } from '@/utils/hooks/useNotification';
-import { ShareModal as CatalogShare } from '@sistent/sistent';
-import { getShareableResourceRoute } from './hooks';
+import { getShareableResourceRoute } from './SpacesSwitcher/hooks';
 import { JsonParse } from '@/utils/utils';
 import { MESHERY_CLOUD_PROD } from '@/constants/endpoints';
 import { RESOURCE_TYPE } from '@/utils/Enum';
 
-export const ShareModal_ = ({ selectedResource, dataName, handleShareModalClose }) => {
+type Resource = {
+  id?: string;
+  name?: string;
+  userId?: string;
+  catalogData?: unknown;
+  patternFile?: string;
+};
+
+type ShareWorkspaceModalInternalProps = {
+  selectedResource: Resource | Resource[];
+  dataName: string;
+  handleShareModalClose: () => void;
+};
+
+export const ShareWorkspaceModal_: FC<ShareWorkspaceModalInternalProps> = ({
+  selectedResource,
+  dataName,
+  handleShareModalClose,
+}) => {
   const firstSelectedResource = Array.isArray(selectedResource)
     ? selectedResource[0]
     : selectedResource;
@@ -40,24 +68,20 @@ export const ShareModal_ = ({ selectedResource, dataName, handleShareModalClose 
   const [handleResourceShare] = useCreateAndRevokeResourceAccessRecordMutation();
   const [updateView] = useUpdateViewVisibilityMutation();
 
-  /**
-   * Determines if the visibility of the selected resource can be edited.
-   * @returns {boolean} - `true` if the visibility can be edited, otherwise `false`.
-   */
-
   const fetchAccessActors = () => {
     if (accessActorsFetched && isOwnerDataFetched) {
       return [ownerData, ...accessActorsInfoOfResource.users];
     }
     return [];
   };
+
   const shareableLink = getShareableResourceRoute(
     dataName,
     firstSelectedResource?.id,
     firstSelectedResource?.name,
   );
 
-  const handleUpdatePatternVisibility = async (value) => {
+  const handleUpdatePatternVisibility = async (value: string) => {
     const res = await updatePatterns({
       body: {
         id: firstSelectedResource?.id,
@@ -72,7 +96,7 @@ export const ShareModal_ = ({ selectedResource, dataName, handleShareModalClose 
     };
   };
 
-  const handleUpdateViewVisibility = async (value) => {
+  const handleUpdateViewVisibilityCall = async (value: string) => {
     const res = await updateView({
       id: firstSelectedResource?.id,
       body: {
@@ -84,15 +108,16 @@ export const ShareModal_ = ({ selectedResource, dataName, handleShareModalClose 
     };
   };
 
-  const handleUpdateVisibility = (updatedVisibility) => {
-    if (dataName == RESOURCE_TYPE.DESIGN) {
+  const handleUpdateVisibility = (updatedVisibility: string) => {
+    if (dataName === RESOURCE_TYPE.DESIGN) {
       return handleUpdatePatternVisibility(updatedVisibility);
     }
-
-    if (dataName == RESOURCE_TYPE.VIEW) {
-      return handleUpdateViewVisibility(updatedVisibility);
+    if (dataName === RESOURCE_TYPE.VIEW) {
+      return handleUpdateViewVisibilityCall(updatedVisibility);
     }
+    return undefined;
   };
+
   return (
     <CatalogShare
       selectedResource={selectedResource}
@@ -112,10 +137,20 @@ export const ShareModal_ = ({ selectedResource, dataName, handleShareModalClose 
   );
 };
 
-const ShareModal = ({ resource, type, handleClose }) => {
+export interface ShareWorkspaceModalProps {
+  resource: Resource | Resource[];
+  type: string;
+  handleClose: () => void;
+}
+
+const ShareWorkspaceModal: FC<ShareWorkspaceModalProps> = ({ resource, type, handleClose }) => {
   return (
-    <ShareModal_ handleShareModalClose={handleClose} selectedResource={resource} dataName={type} />
+    <ShareWorkspaceModal_
+      handleShareModalClose={handleClose}
+      selectedResource={resource}
+      dataName={type}
+    />
   );
 };
 
-export default ShareModal;
+export default ShareWorkspaceModal;
