+++ ui/components/workspaces/SpacesSwitcher/MainDesignsContent.tsx
@@ -28,7 +28,7 @@ import { MenuComponent } from './MenuComponent';
 import { DesignList, GhostContainer, GhostImage, GhostText, LoadingContainer } from './styles';
 import ExportModal from '../../shared/Modal/ExportModal';
 import { RESOURCE_TYPE } from '@/utils/Enum';
-import ShareModal from './ShareModal';
+import ShareModal from '../ShareWorkspaceModal';
 import InfoModal from '../../shared/Modal/Information/InfoModal';
 import { useGetMeshModelsQuery } from '@/rtk-query/meshModel';
 import {
