+++ ui/components/workspaces/SpacesSwitcher/MainViewsContent.tsx
@@ -23,8 +23,8 @@ import { MenuComponent } from './MenuComponent';
 import { RESOURCE_TYPE } from '@/utils/Enum';
 import { DesignList, LoadingContainer, GhostContainer, GhostImage, GhostText } from './styles';
 import { useUpdateViewVisibilityMutation } from '@/rtk-query/view';
-import ShareModal from './ShareModal';
-import { ViewInfoModal } from '../../shared/Modal/ViewInfoModal';
+import ShareModal from '../ShareWorkspaceModal';
+import { ViewInfoModal } from '../ViewInfoModal';
 import { openViewInKanvas, useIsOperatorEnabled } from '@/utils/utils';
 import { useNotification } from '@/utils/hooks/useNotification';
 import { EVENT_TYPES } from 'lib/event-types';
