+++ ui/eslint.config.js
@@ -83,11 +83,9 @@ const legacyRestrictedImportOffenders = [
   'components/Settings/Registry/Stepper/CSVStepper.tsx',
   'components/Settings/Registry/Stepper/UrlStepper.tsx',
   'components/SpacesSwitcher/MainDesignsContent.tsx',
-  'components/SpacesSwitcher/WorkspaceModal.tsx',
   'components/SpacesSwitcher/components.tsx',
   'components/TroubleshootingModalComponent.tsx',
   'components/UserPreferences/index.tsx',
-  'components/ViewInfoModal.tsx',
   'components/ViewSwitch.tsx',
   'components/YamlDialog.tsx',
   'components/configuratorComponents/MeshModel/index.tsx',
@@ -179,7 +177,6 @@ const legacyLiteralColorOffenders = [
   'components/telemetry/grafana/GrafanaCustomGaugeChart.tsx',
   'components/telemetry/grafana/GrafanaDateRangePicker.tsx',
   'css/icons.styles.ts',
-  'pages/extension/AccessMesheryModal.tsx',
   'utils/custom-search.tsx',
 ];
 
@@ -348,13 +345,11 @@ const legacyInlineStyleOffenders = [
   'components/workspaces/SpacesSwitcher/SharedContent.tsx',
   'components/workspaces/SpacesSwitcher/SpaceSwitcher.tsx',
   'components/workspaces/SpacesSwitcher/WorkspaceContent.tsx',
-  'components/workspaces/SpacesSwitcher/WorkspaceModal.tsx',
   'components/workspaces/SpacesSwitcher/WorkspaceSwitcher.tsx',
   'components/workspaces/SpacesSwitcher/components.tsx',
   'components/TroubleshootingComponent.tsx',
   'components/shared/FormFields/typing-filter/index.tsx',
   'components/user-preferences/index.tsx',
-  'components/shared/Modal/ViewInfoModal.tsx',
   'components/ViewSwitch.tsx',
   'components/designs/configurator/MeshModel/LazyComponentForm.tsx',
   'components/designs/configurator/MeshModel/index.tsx',
@@ -385,7 +380,6 @@ const legacyInlineStyleOffenders = [
   'components/telemetry/prometheus/PrometheusSelectionComponent.tsx',
   'pages/_app.tsx',
   'pages/_document.tsx',
-  'pages/extension/AccessMesheryModal.tsx',
   'pages/extensions.tsx',
   'utils/custom-search.tsx',
   'utils/utils.tsx',
